+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/handlerAdapter/HandlerAdapterRegistry.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc.tobe.handlerAdapter;
 
 import java.util.ArrayList;
 import java.util.List;
