+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/handlerMapping/AnnotationHandlerMapping.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc.tobe.handlerMapping;
 
 import java.lang.reflect.Method;
 import java.util.Arrays;
